requirements.txt
flask
numpy
sentence-transformers
torch
python-dotenv
openai
Pillow